+++ src/Runly/ConfigReader.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Reflection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Runly.Hosting;
 
 namespace Runly
 {
