+++ src/Runly/Hosting/Debug.cs
@@ -1,4 +1,4 @@
-﻿namespace Runly.Internal
+﻿namespace Runly.Hosting
 {
 	class Debug
 	{
