+++ src/Runly/Hosting/GetAction.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Runly.Internal
+namespace Runly.Hosting
 {
 	class GetAction : IHostAction
 	{
