+++ src/Runly/JobHost.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Runly.Internal;
+using Runly.Hosting;
 
 namespace Runly
 {
