+++ src/Runly/ResultLog.cs
@@ -1,5 +1,6 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+using Runly.Hosting;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
