+++ src/Runly/Util/PollingStream.cs
