+++ test/Runly.Tests/FakeResultsConnection.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.SignalR.Client;
+using Runly.Client;
 
 namespace Runly.Tests
 {
