+++ test/Runly.Tests/Scenarios/Configuration/Generating_config_schema.cs
@@ -1,5 +1,5 @@
 using FluentAssertions;
-using Runly.Internal;
+using Runly.Hosting;
 using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
